Format code to fix checkstyle error
