+++ src/main/java/eureka/command/DeleteCommand.java
@@ -1,7 +1,7 @@
 package eureka.command;
 
 import eureka.Storage;
-import eureka.Task.TaskList;
+import eureka.task.TaskList;
 import eureka.ui.Ui;
 
 /**
