+++ src/main/java/eureka/task/Event.java
@@ -1,11 +1,23 @@
-package eureka.Task;
+package eureka.task;
 
 import java.time.LocalDateTime;
 
+/**
+ * Represents an event task with start and end times.
+ * An event is a task that occurs during a specific time period.
+ */
 public class Event extends Task {
     private LocalDateTime start;
     private LocalDateTime end;
 
+    /**
+     * Creates a new Event with the specified start time, end time, task name, and completion status.
+     *
+     * @param start the start date and time of the event
+     * @param end the end date and time of the event
+     * @param taskName the name/description of the event
+     * @param isDone whether the event is marked as completed
+     */
     public Event(LocalDateTime start, LocalDateTime end, String taskName, boolean isDone) {
         super(taskName, isDone);
         this.start = start;
@@ -15,7 +27,7 @@ public class Event extends Task {
     @Override
     public String toString() {
         String fromStr = (start != null) ? super.dateTimeFormatter(start) : "N/A";
-        String toStr   = (end   != null) ? super.dateTimeFormatter(end)   : "N/A";
+        String toStr = (end != null) ? super.dateTimeFormatter(end) : "N/A";
 
         return "[E] " + super.toString()
                 + " (from: " + fromStr
